requirements.txt
allure-pytest==2.9.42
allure-python-commons==2.9.42
apipkg==1.5
astor==0.8.1
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
colorlog==5.0.1
configparser==5.0.2
crayons==0.4.0
easydev==0.11.1
et-xmlfile==1.1.0
execnet==1.8.0
idna==2.10
iniconfig==1.1.1
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
openpyxl==3.0.7
packaging==20.9
pexpect==4.8.0
pluggy==0.13.1
ptyprocess==0.7.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-ast-transformer==1.0.3
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-soft-assertions==0.1.2
pytest-xdist==2.2.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
texttable==1.6.3
toml==0.10.2
urllib3==1.26.5
webdriver-manager==3.4.2
